Fix display error during camera init.
A camera can attempt to initialize before OnAppyTemplate(), which gets the error control from the template. If an error occurs during camera initialization (no plugin, etc.) and the error control is not ready at the time the error message is set, we save the pending error until it can be displayed.