Implement move camera using mouse
Add a camera setting that enables a camera that supports motion to be moved by holding down the Control key and dragging the mouse over the camera image. This feature was already part of the camera control (default turned off). These changes enable the user to enable/disable the feature as part of the camera settings.